Moved utility casting functions to shared
