+++ src/systems/a320_systems_wasm/src/lib.rs
@@ -3,6 +3,7 @@ use a320_systems::A320;
 use msfs::sim_connect::{SimConnectRecv, SIMCONNECT_OBJECT_ID_USER};
 use msfs::{legacy::execute_calculator_code, sim_connect::SimConnect, sys};
 use std::{pin::Pin, time::Duration};
+use systems::shared::{f64_to_sim_connect_32k_pos, sim_connect_32k_pos_to_f64};
 use systems_wasm::{
     HandleMessage, MsfsAircraftVariableReader, MsfsNamedVariableReaderWriter,
     MsfsSimulationHandler, PrePostTick, ReadWrite, SimulatorAspect,
@@ -178,28 +179,6 @@ impl ElectricalBusConnection {
     }
 }
 
-const MIN_32KPOS_VAL_FROM_SIMCONNECT: f64 = -16384.;
-const MAX_32KPOS_VAL_FROM_SIMCONNECT: f64 = 16384.;
-const RANGE_32KPOS_VAL_FROM_SIMCONNECT: f64 =
-    MAX_32KPOS_VAL_FROM_SIMCONNECT - MIN_32KPOS_VAL_FROM_SIMCONNECT;
-const OFFSET_32KPOS_VAL_FROM_SIMCONNECT: f64 = 16384.;
-// Takes a 32k position type from simconnect, returns a value from scaled from 0 to 1
-fn sim_connect_32k_pos_to_f64(sim_connect_axis_value: u32) -> f64 {
-    let casted_value = (sim_connect_axis_value as i32) as f64;
-    let scaled_value =
-        (casted_value + OFFSET_32KPOS_VAL_FROM_SIMCONNECT) / RANGE_32KPOS_VAL_FROM_SIMCONNECT;
-
-    scaled_value.min(1.).max(0.)
-}
-// Takes a [0:1] f64 and returns a simconnect 32k position type
-fn f64_to_sim_connect_32k_pos(scaled_axis_value: f64) -> u32 {
-    let back_to_position_format = ((scaled_axis_value) * RANGE_32KPOS_VAL_FROM_SIMCONNECT)
-        - OFFSET_32KPOS_VAL_FROM_SIMCONNECT;
-    let to_i32 = back_to_position_format as i32;
-    let to_u32 = to_i32 as u32;
-
-    to_u32
-}
 struct Brakes {
     id_brake_left: sys::DWORD,
     id_brake_right: sys::DWORD,
